chore(ui): move auth functions into utils folder
