+++ src/main/ui/app/utils/auth.ts
